requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
defusedxml==0.8.0rc2
Django==4.2.7
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-nested-routers==0.93.4
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
idna==3.6
importlib-resources==6.1.1
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.4
matplotlib==3.8.2
mpmath==1.3.0
mysqlclient==2.2.0
networkx==3.2.1
numpy==1.26.3
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
Pillow==10.1.0
psutil==5.9.8
py-cpuinfo==9.0.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
sympy==1.12
thop==0.1.1.post2209072238
torch==2.2.0
torchaudio==2.0.2+cu117
torchvision==0.17.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.1.9
urllib3==2.1.0
zipp==3.17.0
